app/requirements.txt
aiolimiter==1.1.0; python_version >= "3.8" and python_version < "4.0"
anyio==3.7.1; python_version >= "3.8"
apscheduler==3.10.3; python_version >= "3.8"
cachetools==5.3.1; python_version >= "3.8"
certifi==2023.7.22; python_version >= "3.8"
charset-normalizer==3.2.0; python_full_version >= "3.7.0" and python_version >= "3.7"
exceptiongroup==1.1.2; python_version < "3.11" and python_version >= "3.8"
h11==0.14.0; python_version >= "3.8"
httpcore==0.17.3; python_version >= "3.8"
httpx==0.24.1; python_version >= "3.8"
idna==3.4; python_version >= "3.8"
python-telegram-bot==20.4; python_version >= "3.8"
pytz==2023.3; python_version >= "3.8"
requests==2.31.0; python_version >= "3.7"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_version >= "3.8" and python_full_version >= "3.3.0"
sniffio==1.3.0; python_version >= "3.8"
tornado==6.3.3; python_version >= "3.8"
tzdata==2023.3; python_version >= "3.8" and platform_system == "Windows"
tzlocal==5.0.1; python_version >= "3.8"
urllib3==2.0.4; python_version >= "3.7"
